+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### New Features
 
  - add new scope "instruction" for matching mnemonics and operands #767 @williballenthin
- - add new feature "operand[{0, 1, 2}].immediate" for matching instruction operand immediate values #767 @williballenthin
+ - add new feature "operand[{0, 1, 2}].number" for matching instruction operand immediate values #767 @williballenthin
  - add new feature "operand[{0, 1, 2}].offset" for matching instruction operand offsets #767 @williballenthin
 
 ### Breaking Changes